signed_url_service/requirements.txt
Flask==2.3.2
google-cloud-storage==2.11.0
google-auth==2.23.0
google-auth-httplib2==0.1.0
requests==2.31.0
google-cloud-firestore==2.11.0
google-cloud-aiplatform==1.30.0
google-cloud-documentai==2.10.0   # optional
pdfminer.six==20221105
python-docx==0.8.11
requests==2.31.0
gunicorn==20.1.0
